+++ backend/canisters/notifications_index/impl/src/updates/add_notifications_canister.rs
@@ -5,7 +5,6 @@ use canister_tracing_macros::trace;
 use notifications_index_canister::add_notifications_canister::{Response::*, *};
 use notifications_index_canister::{NotificationsIndexEvent, SubscriptionAdded};
 use rand::RngCore;
-use std::collections::hash_map::Entry::Vacant;
 use types::{BuildVersion, CanisterId, CanisterWasm, IdempotentEnvelope};
 use utils::canister::{install_basic, set_controllers};
 
@@ -59,54 +58,62 @@ fn commit(
     wasm_version: BuildVersion,
     state: &mut RuntimeState,
 ) -> Response {
+    let now = state.env.now();
     state.data.local_indexes.insert(local_user_index_canister_id);
+    state
+        .data
+        .notifications_canisters
+        .insert(canister_id, NotificationsCanister::new(wasm_version, now));
 
-    if let Vacant(e) = state.data.notifications_canisters.entry(canister_id) {
-        let now = state.env.now();
-        e.insert(NotificationsCanister::new(wasm_version, now));
+    let mut events = vec![NotificationsIndexEvent::SetNotificationPusherPrincipals(
+        state.data.push_service_principals.clone(),
+    )];
 
-        for (user_id, subscription) in state
-            .data
-            .subscriptions
-            .iter()
-            .flat_map(|(user_id, subs)| subs.iter().map(|s| (*user_id, s.clone())))
-        {
-            state.data.notification_canisters_event_sync_queue.push(
-                canister_id,
-                IdempotentEnvelope {
-                    created_at: now,
-                    idempotency_id: state.env.rng().next_u64(),
-                    value: NotificationsIndexEvent::SubscriptionAdded(SubscriptionAdded { user_id, subscription }),
-                },
-            );
-        }
-
-        for (user_id, blocked_users) in state.data.blocked_users.collect_all() {
-            for blocked_user in blocked_users {
-                state.data.notification_canisters_event_sync_queue.push(
-                    canister_id,
-                    IdempotentEnvelope {
-                        created_at: now,
-                        idempotency_id: state.env.rng().next_u64(),
-                        value: NotificationsIndexEvent::UserBlocked(user_id, blocked_user),
-                    },
-                );
-            }
-        }
-
-        for (user_id, endpoint) in state.data.bot_endpoints.iter() {
-            state.data.notification_canisters_event_sync_queue.push(
-                canister_id,
-                IdempotentEnvelope {
-                    created_at: now,
-                    idempotency_id: state.env.rng().next_u64(),
-                    value: NotificationsIndexEvent::BotEndpointUpdated(*user_id, endpoint.clone()),
-                },
-            );
-        }
-
-        Success
-    } else {
-        AlreadyAdded
+    for (user_id, subscription) in state
+        .data
+        .subscriptions
+        .iter()
+        .flat_map(|(user_id, subs)| subs.iter().map(|s| (*user_id, s.clone())))
+    {
+        events.push(NotificationsIndexEvent::SubscriptionAdded(SubscriptionAdded {
+            user_id,
+            subscription,
+        }));
     }
+
+    state.data.local_index_event_sync_queue.push_many(
+        local_user_index_canister_id,
+        events
+            .iter()
+            .map(|e| IdempotentEnvelope {
+                created_at: now,
+                idempotency_id: state.env.rng().next_u64(),
+                value: e.clone(),
+            })
+            .collect(),
+    );
+
+    for (user_id, blocked_users) in state.data.blocked_users.collect_all() {
+        for blocked_user in blocked_users {
+            events.push(NotificationsIndexEvent::UserBlocked(user_id, blocked_user));
+        }
+    }
+
+    for (user_id, endpoint) in state.data.bot_endpoints.iter() {
+        events.push(NotificationsIndexEvent::BotEndpointUpdated(*user_id, endpoint.clone()));
+    }
+
+    state.data.notification_canisters_event_sync_queue.push_many(
+        canister_id,
+        events
+            .into_iter()
+            .map(|e| IdempotentEnvelope {
+                created_at: now,
+                idempotency_id: state.env.rng().next_u64(),
+                value: e,
+            })
+            .collect(),
+    );
+
+    Success
 }