+++ backend/integration_tests/src/notification_tests.rs
@@ -1,7 +1,6 @@
 use crate::env::ENV;
 use crate::utils::tick_many;
 use crate::{CanisterIds, TestEnv, User, client};
-use candid::Principal;
 use itertools::Itertools;
 use pocket_ic::PocketIc;
 use std::ops::Deref;
@@ -21,15 +20,16 @@ fn direct_message_notification_succeeds() {
 
     let TestData { user1, user2 } = init_test_data(env, canister_ids);
 
-    let notifications_canister = canister_ids.notifications(env, user2.canister());
-    let latest_notification_index = latest_notification_index(env, notifications_canister, *controller);
+    let local_user_index_canister = canister_ids.local_user_index(env, user2.canister());
+    let latest_notification_index =
+        client::local_user_index::happy_path::latest_notification_index(env, *controller, local_user_index_canister);
 
     client::user::happy_path::send_text_message(env, &user1, user2.user_id, random_string(), None);
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 1,
     );
 
@@ -49,8 +49,9 @@ fn group_message_notification_succeeds() {
 
     let TestData { user1, user2 } = init_test_data(env, canister_ids);
 
-    let notifications_canister = canister_ids.notifications(env, user2.canister());
-    let latest_notification_index = latest_notification_index(env, notifications_canister, *controller);
+    let local_user_index_canister = canister_ids.local_user_index(env, user2.canister());
+    let latest_notification_index =
+        client::local_user_index::happy_path::latest_notification_index(env, *controller, local_user_index_canister);
 
     let group_id = client::user::happy_path::create_group(env, &user1, &random_string(), false, false);
     client::local_user_index::happy_path::add_users_to_group(
@@ -63,10 +64,10 @@ fn group_message_notification_succeeds() {
 
     client::group::happy_path::send_text_message(env, &user1, group_id, None, random_string(), None);
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 1,
     );
 
@@ -97,15 +98,16 @@ fn direct_message_notification_muted() {
         },
     );
 
-    let notifications_canister = canister_ids.notifications(env, user2.canister());
-    let latest_notification_index = latest_notification_index(env, notifications_canister, *controller);
+    let local_user_index_canister = canister_ids.local_user_index(env, user2.canister());
+    let latest_notification_index =
+        client::local_user_index::happy_path::latest_notification_index(env, *controller, local_user_index_canister);
 
     client::user::happy_path::send_text_message(env, &user1, user2.user_id, random_string(), None);
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 1,
     );
 
@@ -146,8 +148,9 @@ fn group_message_notification_muted(case: u32) {
         &group_canister::toggle_mute_notifications::Args { mute: true },
     );
 
-    let notifications_canister = canister_ids.notifications(env, user2.canister());
-    let latest_notification_index = latest_notification_index(env, notifications_canister, *controller);
+    let local_user_index_canister = canister_ids.local_user_index(env, user2.canister());
+    let latest_notification_index =
+        client::local_user_index::happy_path::latest_notification_index(env, *controller, local_user_index_canister);
 
     let (text, mentioned) = match case {
         1 => (random_string(), Vec::new()),
@@ -183,10 +186,10 @@ fn group_message_notification_muted(case: u32) {
         },
     );
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 1,
     );
 
@@ -222,15 +225,16 @@ fn only_store_up_to_10_subscriptions_per_user() {
 
     env.tick();
 
-    let notifications_canister = canister_ids.notifications(env, user2.canister());
-    let latest_notification_index = latest_notification_index(env, notifications_canister, *controller);
+    let local_user_index_canister = canister_ids.local_user_index(env, user2.canister());
+    let latest_notification_index =
+        client::local_user_index::happy_path::latest_notification_index(env, *controller, local_user_index_canister);
 
     client::user::happy_path::send_text_message(env, &user1, user2.user_id, random_string(), None);
 
-    let mut notifications_response = client::notifications::happy_path::notifications_v2(
+    let mut notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 1,
     );
 
@@ -263,15 +267,16 @@ fn notifications_blocked_from_blocked_users() {
         vec![(user2.user_id, user2.principal)],
     );
 
-    let notifications_canister = canister_ids.notifications(env, group_id);
-    let latest_notification_index = latest_notification_index(env, notifications_canister, *controller);
+    let local_user_index_canister = canister_ids.local_user_index(env, group_id);
+    let latest_notification_index =
+        client::local_user_index::happy_path::latest_notification_index(env, *controller, local_user_index_canister);
 
     client::group::happy_path::send_text_message(env, &user1, group_id, None, random_string(), None);
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 1,
     );
 
@@ -284,10 +289,10 @@ fn notifications_blocked_from_blocked_users() {
 
     client::group::happy_path::send_text_message(env, &user1, group_id, None, random_string(), None);
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 2,
     );
 
@@ -299,27 +304,15 @@ fn notifications_blocked_from_blocked_users() {
 
     client::group::happy_path::send_text_message(env, &user1, group_id, None, random_string(), None);
 
-    let notifications_response = client::notifications::happy_path::notifications_v2(
+    let notifications_response = client::local_user_index::happy_path::notifications_v2(
         env,
         *controller,
-        notifications_canister,
+        local_user_index_canister,
         latest_notification_index + 2,
     );
 
     assert_eq!(notifications_response.notifications.len(), 1);
     assert!(notifications_response.subscriptions.contains_key(&user2.user_id));
-}
-
-fn latest_notification_index(env: &PocketIc, notifications_canister_id: Principal, controller: Principal) -> u64 {
-    let notifications_canister::latest_notification_index::Response::Success(latest_notification_index) =
-        client::notifications::latest_notification_index(
-            env,
-            controller,
-            notifications_canister_id,
-            &notifications_canister::latest_notification_index::Args {},
-        );
-
-    latest_notification_index
 }
 
 fn init_test_data(env: &mut PocketIc, canister_ids: &CanisterIds) -> TestData {