+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Increase scope of `bot_subscribe_to_events` ([#7996](https://github.com/open-chat-labs/open-chat/pull/7996))
 - Allow bot to specify data encoding format ([#8004](https://github.com/open-chat-labs/open-chat/pull/8004))
 
+### Fixed
+
+- Fix ordering of blocked user map ([#8003](https://github.com/open-chat-labs/open-chat/pull/8003))
+
 ## [[2.0.1748](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1748-local_user_index)] - 2025-05-13
 
 ### Changed